requirements.txt
fastapi
uvicorn
pydantic
networkx
transformers
sympy
pandas
scipy
matplotlib
scikit-learn
sentence-transformers
torch
pyautogen
